More manip

+++ BattleshipGame2/BattleshipGame2/PlayerResult.cpp
@@ -6,9 +6,10 @@ using namespace std;
 PlayerResult::PlayerResult(const std::string& name) : _totalNumPointsFor(0), _totalNumPointsAgainst(0), _totalNumLosses(0), _totalNumWins(0), _name(name)
 {
 }
-int PlayerResult::getWinPercentage() const
+
+float PlayerResult::getWinPercentage() const
 {
-	int totalGames = _totalNumLosses + _totalNumWins;
+	float totalGames = _totalNumLosses + _totalNumWins;
 	if (totalGames == 0)
 	{
 		return 0;
